+++ src/map/npc.h
@@ -41,7 +41,7 @@ enum {
 struct view_data* npc_get_viewdata(int class_);
 int npc_chat_sub(struct block_list* bl, va_list ap);
 int npc_event_dequeue(struct map_session_data* sd);
-int npc_event_timer(int tid,unsigned int tick, int id, int data);
+int npc_event_timer(int tid,unsigned int tick, intptr_t id, intptr_t data);
 int npc_event(struct map_session_data* sd, const char* eventname, int mob_kill);
 int npc_timer_event(const char* eventname); // Added by RoVeRT
 int npc_command(struct map_session_data* sd, const char* npcname, const char* command);
