+++ src/map/unit.c
@@ -45,8 +45,8 @@ struct unit_data* unit_bl2ud(struct block_list *bl)
 	return NULL;
 }
 
-static int unit_attack_timer(int tid,unsigned int tick,int id,int data);
-static int unit_walktoxy_timer(int tid,unsigned int tick,int id,int data);
+static int unit_attack_timer(int tid,unsigned int tick,intptr_t id,intptr_t data);
+static int unit_walktoxy_timer(int tid,unsigned int tick,intptr_t id,intptr_t data);
 
 int unit_walktoxy_sub(struct block_list *bl)
 {
@@ -100,7 +100,7 @@ int unit_walktoxy_sub(struct block_list *bl)
 	return 1;
 }
 
-static int unit_walktoxy_timer(int tid,unsigned int tick,int id,int data)
+static int unit_walktoxy_timer(int tid,unsigned int tick,intptr_t id,intptr_t data)
 {
 	int i;
 	int x,y,dx,dy,dir;
@@ -321,7 +321,7 @@ int unit_walktoxy( struct block_list *bl, int x, int y, int flag)
 	if((bl)->type == BL_MOB && (flag)) \
 		((TBL_MOB*)(bl))->state.skillstate = ((TBL_MOB*)(bl))->state.aggressive?MSS_FOLLOW:MSS_RUSH;
 
-static int unit_walktobl_sub(int tid,unsigned int tick,int id,int data)
+static int unit_walktobl_sub(int tid,unsigned int tick,intptr_t id,intptr_t data)
 {
 	struct block_list *bl = map_id2bl(id);
 	struct unit_data *ud = bl?unit_bl2ud(bl):NULL;
@@ -1159,7 +1159,7 @@ int unit_skilluse_pos2( struct block_list *src, int skill_x, int skill_y, int sk
 	return 1;
 }
 
-static int unit_attack_timer(int tid,unsigned int tick,int id,int data);
+static int unit_attack_timer(int tid,unsigned int tick,intptr_t id,intptr_t data);
 
 int unit_stop_attack(struct block_list *bl)
 {
@@ -1449,7 +1449,7 @@ static int unit_attack_timer_sub(struct block_list* src, int tid, unsigned int t
 	return 1;
 }
 
-static int unit_attack_timer(int tid,unsigned int tick,int id,int data)
+static int unit_attack_timer(int tid,unsigned int tick,intptr_t id,intptr_t data)
 {
 	struct block_list *bl;
 	bl = map_id2bl(id);
